wip: add security provider interface (SPI)
The motivation for this is to extract the security/crypto functions
specific to OpenSSL into an interface allowing other implementation.

The SPI will not have any dependenies on V8. The idea currently is to
extract functions from node_crypto.cc and leave the V8 specific code
in node_crypto.cc which will use the SPI to call security/crypto
functions.