v8: remove void* cast in ptr_arg_cmd
This commit removes the addition of a void* cast ptr_arg_cmd. The
motivation of this is that currently the jlh command does not work. For
example:
(lldb) jlh env()->quic_on_socket_ready_function()

With this change the jlh command works as expected:

(lldb) (lldb) jlh env()->quic_on_socket_ready_function()
0x3d6bd6c81c29: [Function]
 - map: 0x3d6babdc0751 <Map(HOLEY_ELEMENTS)> [FastProperties]
 - prototype: 0x3d6b6f140979 <JSFunction (sfi = 0x3d6ba9988179)>
 - elements: 0x3d6bbfe00bf9 <FixedArray[0]> [HOLEY_ELEMENTS]
 - function prototype:
 - initial_map:
 - shared_info: 0x3d6b45e48011 <SharedFunctionInfo onSocketReady>
 - name: 0x3d6b45e48069 <String[#13]: onSocketReady>
 - builtin: CompileLazy
 - formal_parameter_count: 1
 - kind: NormalFunction
 - context: 0x3d6bd6ca1e51 <FunctionContext[144]>
 - code: 0x1c1635244721 <Code BUILTIN CompileLazy>
 - source code: (fd) {
  this[owner_symbol][kReady](fd);
}
 - properties: 0x3d6bbfe00bf9 <FixedArray[0]> {
    #length: 0x3d6ba9980491 <AccessorInfo> (const accessor descriptor)
    #name: 0x3d6ba9980421 <AccessorInfo> (const accessor descriptor)
    #prototype: 0x3d6ba9980501 <AccessorInfo> (const accessor descriptor)
 }

 - feedback vector: feedback metadata is not available in SFI

I'm not sure yet how the other commands that use ptr_arg_cmd are
affected but thought I'd bring this up as a WIP incase others are seeing
the same issue.