tools,build: add PRODUCT_DIR_ABS default gyp var
This commit adds a default variable named PRODUCT_DIR_ABS which contains
the absolute path to the configured output and buildtype (Release/Debug)
directory.

The motivation for adding this instead of using PRODUCT_DIR is that the
value of PRODUCT_DIR is a ${builddir} variable which is fine in most
situations. The use case we ran into was where we wanted to add a
defines for OpenSSL 3.0 MODULESDIR macro. I this case we want the value
would ${builddir} would be escaped into $${builddir} which would not be
resolved when passed with -DMODULEDIR, so the actual value would
contain the variable ${buildtype} instead of an actual path.